pyproject.toml
[project]
name = "spike-docdb-blowout"
version = "0.1.0"
description = "Spike for comparing various different document store databases."
authors = [
    {name = "Bheesham Persaud",email = "bhee@mozilla.com"}
]
license = {text = "MPL-2.0"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.115.6,<0.116.0)",
    "pymongo (>=4.10.1,<5.0.0)",
    "google-cloud-bigtable (>=2.28.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.14.1"
black = "^24.10.0"
pylint = "^3.3.3"

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 120
disable = """
    missing-module-docstring,
    missing-class-docstring,
    missing-function-docstring,
    invalid-name,
    wildcard-import,
    unused-wildcard-import
"""
